Expand the script to allow changing statuses
