+++ package.json
@@ -77,8 +77,8 @@
     "addMissingGrantsToReports:local": "./node_modules/.bin/babel-node ./src/tools/addMissingGrantsToReportsCLI.js",
     "processData": "node ./build/server/tools/processDataCLI.js",
     "processData:local": "./node_modules/.bin/babel-node ./src/tools/processDataCLI.js",
-    "deleteApproved": "node ./build/server/tools/deleteApprovedCLI.js",
-    "deleteApproved:local": "./node_modules/.bin/babel-node ./src/tools/deleteApprovedCLI.js"
+    "changeReportStatus": "node ./build/server/tools/changeReportStatusCLI.js",
+    "changeReportStatus:local": "./node_modules/.bin/babel-node ./src/tools/changeReportStatusCLI.js"
   },
   "repository": {
     "type": "git",
